+++ src/main/java/Model/Flash.java
@@ -3,7 +3,7 @@ package Model;
 public class Flash extends Digital {
     private String usbVersion;
 
-    public Flash(String productName,long productPrice,double averageScore,int capacity,category product,double weight, double dimension , String usbVersion,String comment) {
+    public Flash(String productName,double productPrice,double averageScore,int capacity,category product,double weight, double dimension , String usbVersion,String comment) {
         super(productName,productPrice, averageScore,capacity,product,weight,dimension,comment);
         this.usbVersion=usbVersion;
     }
