+++ src/main/java/Model/bikeCycle.java
@@ -7,7 +7,7 @@ public class bikeCycle extends Vehicle {
 
     public bikeCycle1 bike;
 
-    public bikeCycle(String productName, long productPrice, double averageScore, int capacity, category product, String company, bikeCycle1 bike,String comment) {
+    public bikeCycle(String productName, double productPrice, double averageScore, int capacity, category product, String company, bikeCycle1 bike,String comment) {
         super(productName,productPrice, averageScore,capacity,product,company,comment);
         this.bike = bike;
     }
